requirements.txt
fastapi
uvicorn
python-dotenv
groq

llama-index-core
llama-index-readers-file
llama-index-vector-stores-qdrant
llama-index-embeddings-huggingface

qdrant-client

torch==2.4.1
transformers==4.44.2
sentence-transformers==3.0.1
